Chinese CVV: custom dictionary support
